mouse position must be a float

+++ mcio_remote/network.py
@@ -92,7 +92,7 @@ class ActionPacket:
     mouse_buttons: List[Tuple[int, int]] = field(default_factory=list)   # List of (button, action) pairs
 
     # List of (x, y) pairs. Using a list for consistency
-    mouse_pos: List[Tuple[int, int]] = field(default_factory=list)
+    mouse_pos: List[Tuple[float, float]] = field(default_factory=list)
 
     def pack(self) -> bytes:
         pkt_dict = asdict(self)
